# Conflicts:
#	backend/build.gradle
#	backend/src/main/java/com/team5/secondhand/api/chatroom/domian/Chatroom.java
#	backend/src/main/java/com/team5/secondhand/api/chatroom/repository/ChatroomRepository.java
#	backend/src/main/java/com/team5/secondhand/api/item/controller/v1/ItemControllerV1.java
#	backend/src/main/java/com/team5/secondhand/api/item/controller/v2/ItemControllerV2.java
#	backend/src/main/java/com/team5/secondhand/api/item/controller/v2/ItemFacade.java
#	backend/src/main/java/com/team5/secondhand/api/item/domain/Item.java
#	backend/src/main/java/com/team5/secondhand/api/resource/controller/ResourceController.java
#	backend/src/main/java/com/team5/secondhand/api/resource/service/ResourceService.java
#	backend/src/main/java/com/team5/secondhand/chat/chatroom/config/ChatroomConfig.java
#	backend/src/main/java/com/team5/secondhand/global/config/RedisConfig.java
#	backend/src/main/resources/application.yml

+++ backend/src/main/java/com/team5/secondhand/api/item/service/ItemPostService.java
@@ -1,6 +1,6 @@
 package com.team5.secondhand.api.item.service;
 
-import com.team5.secondhand.api.item.controller.v1.dto.request.ItemPostWithUrl;
+import com.team5.secondhand.api.item.controller.v1.dto.request.ItemUpdateRequest;
 import com.team5.secondhand.api.item.domain.Item;
 import com.team5.secondhand.api.item.domain.Status;
 import com.team5.secondhand.api.item.exception.ExistItemException;
@@ -10,9 +10,10 @@ import com.team5.secondhand.api.member.exception.UnauthorizedException;
 import com.team5.secondhand.api.region.domain.Region;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.security.sasl.AuthenticationException;
 
 @Service
 @RequiredArgsConstructor
@@ -20,32 +21,68 @@ public class ItemPostService {
 
     private final ItemRepository itemRepository;
 
+    /**
+     * 중고거래 물품을 등록하는 서비스입니다.
+     * @param item 중고거래 물품
+     * @param seller 판매자
+     * @param region 판매 지역
+     * @param thumbnailUrl 섬네일 사진 URL
+     * @return 등록된 중고거래 물품의 primary key
+     */
     @Transactional
     public Long postItem(Item item, Member seller, Region region, String thumbnailUrl) {
-        item.updateThumbnail(thumbnailUrl);
-        itemRepository.save(item.owned(seller, region));
-        return item.getId();
+        Item updatedItem = item.updateThumbnail(thumbnailUrl)
+                .assignOwnership(seller, region);
+        return itemRepository.save(updatedItem).getId();
     }
 
+    /**
+     * 중고거래 물품 거래 글을 수정합니다.
+     * @param id 중고거래물품 primary key
+     * @param request 업데이트할 내용이 있는 DTO
+     * @param seller 중고거래 물품 거래 글 게시자
+     * @throws ExistItemException
+     * @throws UnauthorizedException
+     */
     @Transactional
-    public void updateItem(Long id, ItemPostWithUrl itemPost, Member seller) throws ExistItemException, UnauthorizedException {
-        Item item = itemRepository.findById(id).orElseThrow(() -> new ExistItemException("없는 아이템입니다."));
+    public void updateItem(Long id, ItemUpdateRequest request, Member seller) throws ExistItemException, UnauthorizedException {
+        Item item = itemRepository.findById(id)
+                .orElseThrow(() -> new ExistItemException("없는 아이템입니다."));
 
-        if (!item.isSeller(seller.getId())) {
+        if (!item.isSeller(seller)) {
             throw new UnauthorizedException("본인의 글만 수정할 수 있습니다.");
         }
 
-        Item newItem = item.updatePost(itemPost, itemPost.getImages().get(0).getUrl());
-        itemRepository.save(newItem);
+        Item updatedItem = item.update(request.toEntity());
+        itemRepository.save(updatedItem);
     }
 
+    /**
+     * 중고거래 물품 글의 판매상태를 변경합니다.
+     * @param id 중고거래 물품 글의 primary key
+     * @param status 변경하고 싶은 판매상태
+     * @return 실행 결과
+     */
     @Transactional
-    public boolean updateItemStatus(Long id, Status status) {
+    public boolean updateItemStatus(Long id, Status status, Long memberId) throws AuthenticationException {
+        if (!itemRepository.existsByIdAndSellerId(id, memberId)) {
+            throw new AuthenticationException("글 작성자가 아닙니다.");
+        }
+
         return itemRepository.updateStatus(id, status) == 1;
     }
 
+    /**
+     * 중고거래 물품 글을 삭제합니다.
+     * @param id
+     */
     @CacheEvict(value = "menu", allEntries = true)
-    public void deleteById(Long id) {
+    @Transactional
+    public void deleteById(Long id, Long memberId) throws AuthenticationException {
+        if (!itemRepository.existsByIdAndSellerId(id, memberId)) {
+            throw new AuthenticationException("글 작성자가 아닙니다.");
+        }
+
         itemRepository.deleteById(id);
     }
 }
